+++ android/app/src/main/java/com/on/turip/ui/compose/designsystem/component/ErrorScreen.kt
@@ -23,7 +23,7 @@ import com.on.turip.R
 import com.on.turip.ui.common.error.ErrorUiModel
 import com.on.turip.ui.common.error.ErrorUiState
 import com.on.turip.ui.common.error.toUiModel
-import com.on.turip.ui.compose.theme.TuripTypography
+import com.on.turip.ui.compose.designsystem.theme.TuripTypography
 
 @Composable
 fun ErrorScreen(
