+++ android/app/src/main/java/com/on/turip/ui/compose/login/component/GoogleLoginButton.kt
@@ -23,8 +23,8 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.on.turip.R
-import com.on.turip.ui.compose.theme.TuripTheme
-import com.on.turip.ui.compose.theme.TuripTypography
+import com.on.turip.ui.compose.designsystem.theme.TuripTheme
+import com.on.turip.ui.compose.designsystem.theme.TuripTypography
 
 @Composable
 fun GoogleLoginButton(
