+++ android/app/src/main/java/com/on/turip/ui/compose/setting/SettingScreen.kt
@@ -33,10 +33,10 @@ import com.on.turip.ui.common.error.toUiModel
 import com.on.turip.ui.compose.designsystem.component.TuripAppBar
 import com.on.turip.ui.compose.designsystem.component.TuripDialog
 import com.on.turip.ui.compose.designsystem.component.TuripSnackbar
+import com.on.turip.ui.compose.designsystem.theme.TuripTheme
 import com.on.turip.ui.compose.setting.component.SettingItem
 import com.on.turip.ui.compose.setting.model.SettingUiEffect
 import com.on.turip.ui.compose.setting.model.SettingUiState
-import com.on.turip.ui.compose.theme.TuripTheme
 import timber.log.Timber
 
 @Composable
