+++ android/app/src/main/java/com/on/turip/ui/login/LoginActivity.kt
@@ -7,8 +7,8 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import com.on.turip.data.login.datasource.GoogleCredentialManager
+import com.on.turip.ui.compose.designsystem.theme.TuripTheme
 import com.on.turip.ui.compose.login.LoginScreen
-import com.on.turip.ui.compose.theme.TuripTheme
 import com.on.turip.ui.main.MainActivity
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
