+++ android/app/src/main/java/com/on/turip/ui/setting/SettingActivity.kt
@@ -9,8 +9,8 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.net.toUri
 import com.on.turip.domain.setting.InquiryMail
+import com.on.turip.ui.compose.designsystem.theme.TuripTheme
 import com.on.turip.ui.compose.setting.SettingScreen
-import com.on.turip.ui.compose.theme.TuripTheme
 import com.on.turip.ui.login.LoginActivity
 import dagger.hilt.android.AndroidEntryPoint
